feat: 코인 과거 가격 바인딩 #16
- 현재가 형식 변경
