+++ Huhoe/Huhoe/Domain/UseCases/CoinDetailUseCase.swift
@@ -26,4 +26,8 @@ extension CoinDetailUseCase {
     func fetchTransactionWebSocket(with coinSymbols: [String]) -> Observable<RealTimeCoinPrice> {
         return transactionWebSocketRepository.fetchTransaction(with: coinSymbols)
     }
+    
+    func fetchCoinPriceHistory(with coinSymbol: String) -> Observable<CoinPriceHistory> {
+        return candlestickRepository.fetchCoinPriceHistory(with: coinSymbol)
+    }
 }
